Added a bunch of classes needed for the game
Server.java has been modified to store client threads in an ArrayList
Also, reader and writer instances are created within a thread context

Adding a command line parser file as-is from jargs project.

Added a bunch of classes needed for this game. I think this should be reasonable
enough for the skeletal milestone. We are able to pass messages between server and client
and vice-versa. I haven't yet gotten to creating Message class yet. I'll give it a
shot tomorrow.